refactor: useBannerStyle 적용
